move dns to http
