+++ setup.cfg
@@ -25,7 +25,6 @@ console_scripts =
 mite_stats =
     mite = mite.stats:_MITE_STATS
     mite_http = mite_http:_MITE_STATS
-    mite_http = mite_http:_MITE_DNS_STATS
     mite_selenium = mite_selenium:_MITE_STATS
 
 [options.packages.find]
